[feat] #2 ChargeBox 파일 생성
